Scheduked Remove -implemented. Fine tuning on UI Bug fixes
--HG--
branch : ISSUE-2

+++ src/main/java/com/mnewservice/mcontent/domain/DeliveryPipe.java
@@ -15,7 +15,8 @@ public class DeliveryPipe {
     private String theme;
     private Collection<User> providers;
 
-    private int contentCount;
+    // DTO read-only local variable
+    private long contentCount;
 
     public Long getId() {
         return id;
@@ -57,11 +58,11 @@ public class DeliveryPipe {
         this.providers = providers;
     }
 
-    public int getContentCount() {
+    public long getContentCount() {
         return contentCount;
     }
 
-    public void setContentCount(int contentCount) {
+    public void setContentCount(long contentCount) {
         this.contentCount = contentCount;
     }
